Implement different connection handling
We will now wait for a connection to the network to complete before:
 * emiting events
 * resolving the connect promise

This is handled by waiting for the first unicast packet from the master
to arrive. This system will also be selected as our server / master
system for subsequent requests (e.g. for track metadata)

When a timeout happens during the connect all sockets and connections
will be cleared and closed.